refactor: dto이동, url정리
