+++ src/main/resources/templates/room-update.html
@@ -25,7 +25,7 @@
 <div class="container-fluid booking py-5">
     <div class="container py-5">
         <h1 class="text-white mb-3">객실 정보</h1>
-        <form id="registerForm" method="post" th:action="@{/room/{roomId}/roomUpdatePage(roomId=${room.getRoomId()})}">
+        <form id="registerForm" method="post" th:action="@{/room/{roomId}/roomUpdate(roomId=${room.getRoomId()})}">
             <div class="rooms row g-3">
                 <div class="room row g-3">
                     <h2 class="text-white mb-3">객실 정보</h2>
